Ignore timezone when adjusting by units smaller than a day. Undo scientific time toggle that enables the same, but manually. Fix issue #8.
Scientific time has a well defined unit, viz. a second, and a timezone with constant day length, viz. UTC. Duration of a second, and also of a minute and an hour, is independent of the timezone. So adjusting by these units always gives a “scientific” result; and if one wants to adjust by larger units, treating them as constant in a timezone where they are not constant, they should either switch to UTC, or convert the adjustement to constant units (e.g. add 24 hours to adjust by one astronomical day).